Move example from index2 to index
